refactor(replay): introduce event reader port
